medical-transcription/requirements-worker.txt
# Medical Transcription Worker Dependencies - Minimal
# Core dependencies
redis>=5.0.0
sqlalchemy>=2.0.0
boto3>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0

# BullMQ Python worker (no official client, we'll use redis directly)
# Using taskiq-redis as it provides BullMQ compatible worker pattern
taskiq[redis]>=0.11.0

# Audio processing
# Note: For CUDA support, install torch separately:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
faster-whisper>=1.0.0
torch>=2.1.0

# Audio utilities
numpy>=1.24.0
soundfile>=0.12.0  # For audio file handling
librosa>=0.10.0    # For audio preprocessing
